explicit int cast
git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/loggertools/trunk@361 fae35197-53e2-0310-b5ce-cc51903860e9

+++ src/tp-cenfis-reader.cc
@@ -82,7 +82,7 @@ static const T parseAngle(char **pp) {
 
     *pp = endptr + 1;
 
-    return T(sign * (((n1 * 60) + n2) * 1000 + n3));
+    return T((int)(sign * (((n1 * 60) + n2) * 1000 + n3)));
 }
 
 template<class T, char minusLetter, char plusLetter>
